+++ src/react/loom-app/option-bar/search-bar/styles.css
@@ -1,0 +1,3 @@
+.dataloom-search-bar input {
+    max-width: 200px;
+}
